extract ContextM eval into a module;
